verify format tag: refactor and merge tag function
gpg_verify_tag in tag.c is now called verify_format_tag and takes
an additional format parameter to be passed by both
`git tag -v` or `git verify-tag` if the user intends to silence the
normal GPG tag verification output and print the formatted tag
object instead.
